+++ src/Twilio/Rest/Api/V2010/Account/MessageOptions.cs
@@ -294,10 +294,10 @@ namespace Twilio.Rest.Api.V2010.Account
         ///<summary> The SID of the [Account](https://www.twilio.com/docs/iam/api/account) associated with the Message resources. </summary> 
         public string PathAccountSid { get; set; }
 
-        ///<summary> Filter by recipient. For example: Set this `to` parameter to `+15558881111` to retrieve a list of Message resources with `to` properties of `+15558881111` </summary> 
+        ///<summary> Filter by recipient. For example: Set this parameter to `+15558881111` to retrieve a list of Message resources sent to `+15558881111`. </summary> 
         public Types.PhoneNumber To { get; set; }
 
-        ///<summary> Filter by sender. For example: Set this `from` parameter to `+15552229999` to retrieve a list of Message resources with `from` properties of `+15552229999` </summary> 
+        ///<summary> Filter by sender. For example: Set this parameter to `+15552229999` to retrieve a list of Message resources sent by `+15552229999`. </summary> 
         public Types.PhoneNumber From { get; set; }
 
         ///<summary> Filter by Message `sent_date`. Accepts GMT dates in the following formats: `YYYY-MM-DD` (to find Messages with a specific `sent_date`), `<=YYYY-MM-DD` (to find Messages with `sent_date`s on and before a specific date), and `>=YYYY-MM-DD` (to find Messages with `sent_dates` on and after a specific date). </summary>